% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_palettes.R
\name{usyd_cols}
\alias{usyd_cols}
\title{Function to call hex codes from colourList}
\usage{
usyd_cols(...)
}
\arguments{
\item{...}{a list of colour names from colourList}
}
\value{
A list of hex codes corresponding to the chosen colours
}
\description{
Function to call hex codes from colourList
}
