% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_palettes.R
\name{usyd_pal_gen}
\alias{usyd_pal_gen}
\title{Generates new palettes based on colour selections}
\usage{
usyd_pal_gen(...)
}
\arguments{
\item{...}{Colour names from swatchplot (refer to documentation)}
}
\value{
A colour palette ready for use.
}
\description{
Generates new palettes based on colour selections
}
\examples{
my_palette <- usyd_pal_gen("SecondaryDarkSeafoam", "SecondaryLightSeafoam", "AccentYellow")
}
