% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_palettes.R
\name{usyd_palette}
\alias{usyd_palette}
\title{Return hex codes based on predefined palettes}
\usage{
usyd_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of the desired palette (refer to documentation for options). All colours are from the University of Sydney Brand Guidelines document version 2.0.}

\item{n}{The number of colours required. Most palettes support between 2 - 6 colours. The largest palette is \code{extended}, which contains 21 colours.}

\item{type}{"Discrete" or "continuous". Discrete is recommended for most palettes, unless the number of colours required is less than the number of variables, in which case continuous may be selected.}
}
\description{
Return hex codes based on predefined palettes
}
\examples{
usyd_palette('primary')
usyd_palette('jacaranda')
usyd_palette('extended', 20, type = 'continuous')
}
