% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_palettes.R
\docType{data}
\name{usyd_palettes}
\alias{usyd_palettes}
\title{Complete list of USYD colour palettes}
\format{
An object of class \code{list} of length 20.

An object of class \code{list} of length 20.
}
\usage{
usyd_palettes

usyd_palettes
}
\description{
Complete list of USYD colour palettes

Use purrr to remove the names of usyd_palettes
}
\keyword{datasets}
