% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_palettes.R
\docType{data}
\name{usydSwatchplot}
\alias{usydSwatchplot}
\title{University branding colours}
\format{
An object of class \code{matrix} with 23 rows and 1 columns.
}
\usage{
usydSwatchplot
}
\description{
University branding colours
}
\keyword{datasets}
