% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{util_visualise_pal}
\alias{util_visualise_pal}
\title{Visualise a palette}
\usage{
util_visualise_pal(pal, num = length(pal))
}
\arguments{
\item{pal}{Palette name}
}
\description{
Visualise a palette
}
\examples{
util_visualise_pal(usyd_palettes[["ochre"]])
}
